"Essentially, the experiment of trying to improve compilation times, by reducing the code being generated is a failure.
What do we do next?

I am going to have to thing about it.

We'll probably reduce the dependency on the union types.

The cards, for example, instead of being union types, they can be bare integers.

Ok, let's give that a try.

We'll make a commit here and then replace the unit types in the front end first."