This is the best we can do with regard to GADTs. I don't like how now in hovers we have to check whether it has substituted tvars and then do term substs. That will slow down the inferencer significantly.
Just having the feature here impacts all the others negatively.

But it's just performance and the type inferencer might be fast enough anyway.

Let's move on to testing the compilation of GADTs. Unless there are bugs that need fixing, as far as I am concerned, GADTs work perfectly fine in the top down segment.