2/15/2024
10:15am. Tortured morning. I got up 4h ago and couldn't get back to sleep. A pretty insiduous nightmare awoke me. Now I am stressed as hell. So I might as well do some programming. This kind of tension really brings me back.

You can't make a badass on the backbone of a pussy, so I can only be great while doing programming.

Even if I had the choice to be a baddass, that leads into the direction of leaving away emotion and losing what is important.

So since I have to be a bitch either way, I might as well stop caring about dignity and start caring about my programming.

Let me do it. A single all-powerful program to break all of reality. Another step towards it should be made today.

It won't be complete today, and it won't be complete tomorrow, but what is good is to work towards it.

At some level everyone is a bitch, whether it be towards reality or God. Its pretty insiduous how modern religious teaches that you cannot escape God's grasp even through death. But unfortunately, if God is really interested in you, he can always simulate an universal thread where you don't die. So it's probably right on some level.

In the end, the only choice is to keep going forward in opposition.

10:30am. Programming in VS is kind of a grind. Let me try opening the project in VS Code.

Horray, it actually loads in Ionide. Wonderful.

Now I can use the editor shortcuts I am used to.

I used to like VS, but now that I am used to VSC, it really feels like crap.

11:25am. I am not in exactly the right mind to be programming on the backend.

I wish that I got good sleep.

Right now my inspiration level is very low and this latest attempt at factoring out the backend failed.

I am going to have to copy paste it, implement it naively and then factor things out instead.

My brain isn't really listening to me right now.

But proper techinque eliminates the need to think. It's much easier to abstract patterns that exist than to anticipate abstractions ahead of time.

This shouldn't be so hard.