This should be good enough to extend the language. I really wish that VS Code didn't have those false type errors in the `StartUp.fs` file. Also it doesn't seem it's loading the package references properly. Sigh.
But I am really enjoying writing F# in VS Code.

I am doing some late day programming on my own project since I feel like it. It's an anwer to something.

What I have to do next is modify the backend join points. I'll also have to change the ML library.

In the end, the changes will be significant enough to warrant bumping up a new minor version. Hope it doesn't break anything for i574.

As a reaction to disatisfaction, just working on your own project like this is the answer. It was always the answer.

Thinking back to my school days, just what did the games give me but betrayal in the en?. And after the school ended, what did the trading give me?

I started this journey in pursuit of power and resolution, but perhaps the projects were always the answers unto themselves rather than obstacles to be overcome.

I'll stop here for the day. Tomorrow, paid work starts anew.